ai-service/app/requirements.txt
fastapi==0.112.*
uvicorn==0.30.*
gunicorn==22.*
python-multipart
pillow
# --- real mode deps ---
torch>=2.2
torchvision>=0.17
transformers
sentence-transformers
faiss-cpu
openai>=1.0.0
